requirements.txt
aniso8601==9.0.1
api==0.0.7
async-timeout==4.0.2
autopep8==1.6.0
beautifulsoup4==4.11.1
boto3==1.21.45
botocore==1.24.45
certifi==2021.10.8
charset-normalizer==2.0.12
click==8.1.2
Deprecated==1.2.13
Flask==2.1.1
Flask-RESTful==0.3.9
idna==3.3
importlib-metadata==4.11.3
itsdangerous==2.1.2
Jinja2==3.1.1
jmespath==1.0.0
lyricsgenius==3.0.1
MarkupSafe==2.1.1
nose==1.3.7
packaging==21.3
pycodestyle==2.8.0
pyparsing==3.0.8
python-dateutil==2.8.2
python-dotenv==0.20.0
pytz==2022.1
redis==4.2.2
requests==2.27.1
s3transfer==0.5.2
six==1.16.0
soupsieve==2.3.2.post1
toml==0.10.2
urllib3==1.26.9
Werkzeug==2.1.1
wrapt==1.14.0
zipp==3.8.0
